bismillah yg ini jadi/requirements.txt
cryptography==42.0.5
pyodbc==5.1.0
tkinter==0.1.0
Pillow==10.3.0